refactor(auth): reorganizing package structure for authentication components
