+++ src/main/java/soat/project/fastfoodsoat/infrastructure/client/controller/ClientControllerImpl.java
@@ -10,7 +10,7 @@ import soat.project.fastfoodsoat.infrastructure.client.model.request.CreateClien
 import soat.project.fastfoodsoat.infrastructure.client.model.response.ClientAuthResponse;
 import soat.project.fastfoodsoat.infrastructure.client.model.response.CreateClientResponse;
 import soat.project.fastfoodsoat.infrastructure.client.presenter.ClientPresenter;
-import soat.project.fastfoodsoat.infrastructure.web.presenter.AuthClientPresenter;
+import soat.project.fastfoodsoat.infrastructure.auth.presenter.AuthClientPresenter;
 
 @Component
 public class ClientControllerImpl implements ClientController {
