+++ src/main/java/soat/project/fastfoodsoat/infrastructure/web/rest/api/AuthStaffAPI.java
@@ -12,8 +12,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import soat.project.fastfoodsoat.infrastructure.web.model.request.auth.AuthStaffRequest;
-import soat.project.fastfoodsoat.infrastructure.web.model.response.auth.AuthStaffResponse;
+import soat.project.fastfoodsoat.infrastructure.auth.model.request.AuthStaffRequest;
+import soat.project.fastfoodsoat.infrastructure.auth.model.response.AuthStaffResponse;
 import soat.project.fastfoodsoat.infrastructure.web.model.DefaultApiError;
 
 @Tag(name = "Auths")
